refactor(cli): modularize roxy_cli with idiomatic Rust patterns
- Add #[macro_use] extern crate tracing across workspace for cleaner macro usage
- Extract monolithic lib.rs (585 lines) into 8 focused modules
- Introduce factory types (BackendFactory, GroupFactory, RouterFactory, etc.)
- Add AppBuilder with fluent API following ServerBuilder pattern
- Preserve all existing public APIs as convenience wrappers
